+++ backend/mssaga/src/main/java/mssaga/mssaga/DTO/ClienteDTO.java
@@ -5,7 +5,7 @@ public class ClienteDTO {
     private String nome;
     private String cpf;
     private String email;
-    private float saldo_milhas;
+    private Double saldo_milhas;
     private EnderecoDTO endereco;
 
     public Long getCodigo() {
@@ -49,11 +49,11 @@ public class ClienteDTO {
         this.endereco = endereco;
     }
 
-    public float getSaldo_milhas() {
+    public Double getSaldo_milhas() {
         return saldo_milhas;
     }
 
-    public void setSaldo_milhas(float saldo_milhas) {
+    public void setSaldo_milhas(Double saldo_milhas) {
         this.saldo_milhas = saldo_milhas;
     }
